Add stream conformance tests for BufferedStream
Specifically when used in a connected fashion, wrapped around some other connected stream.
